# docs/swagger.yaml
basePath: /api
definitions:
...
  rip_internal_app_dto.RouteInfo:
    properties:
      description:
        type: string
      distance:
        type: integer
      image_url:
        type: string
      route_id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  rip_internal_app_dto.RouteSpeedRequestInfo:
    properties:
      arrival_date:
        type: string
      ship_speed:
        type: integer
    type: object
...
  rip_internal_app_dto.SpeedRequest:
    properties:
      completion_date:
        type: string
      creation_date:
        type: string
      creator_login:
        type: string
      departure_date:
        type: string
      formation_date:
        type: string
      id:
        type: integer
      moderator_login:
        type: string
      status:
        type: string
    type: object
  rip_internal_app_dto.SpeedRequestDetailedResponse:
    properties:
      result:
        type: integer
      route_req:
        items:
          $ref: '#/definitions/rip_internal_app_dto.RouteSpeedRequestInfo'
        type: array
      routes:
        items:
          $ref: '#/definitions/rip_internal_app_dto.RouteInfo'
        type: array
      speed_request:
        $ref: '#/definitions/rip_internal_app_dto.SpeedRequest'
    type: object
  rip_internal_app_dto.UpdateSpeedRequest:
    properties:
      departure_date:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Route Information Platform API
  title: RIP API
  version: "1.0"
paths:
...
  /speedrequests:
    get:
      description: Get speed requests list. For authentificated users only.
      parameters:
      - description: Date From
        format: date
        in: query
        name: date_from
        type: string
      - description: Date To
        in: query
        name: date_to
        type: string
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: response
          schema:
            items:
              $ref: '#/definitions/rip_internal_app_dto.SpeedRequest'
            type: array
      security:
      - BearerAuth: []
      summary: Get speed requests list
      tags:
      - speedrequests
  /speedrequests/{speed_request_id}:
    delete:
      description: Delete speed request. For authentificated users only.
      parameters:
      - description: Speed Request ID
        in: path
        name: speed_request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete speed request
      tags:
      - speedrequests
    get:
      description: Get speed request by ID. For authentificated users only. Client
        can see only their speed requests
      parameters:
      - description: Speed Request ID
        in: path
        name: speed_request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/rip_internal_app_dto.SpeedRequestDetailedResponse'
      security:
      - BearerAuth: []
      summary: Get speed request by ID
      tags:
      - speedrequests
    put:
      description: Update speed request. For authentificated users only.
      parameters:
      - description: Speed Request ID
        in: path
        name: speed_request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: request
          schema:
            $ref: '#/definitions/rip_internal_app_dto.UpdateSpeedRequest'
      security:
      - BearerAuth: []
      summary: Update speed request
      tags:
      - speedrequests
  /speedrequests/{speed_request_id}/complete:
    put:
      consumes:
      - application/json
      description: Complete speed requests. For moderators only.
      parameters:
      - description: Speed Request ID
        in: path
        name: speed_request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Complete speed request
      tags:
      - speedrequests
  /speedrequests/{speed_request_id}/submit:
    put:
      description: Submit speed request. For authentificated users only.
      parameters:
      - description: Speed Request ID
        in: path
        name: speed_request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/rip_internal_app_dto.SpeedRequestDetailedResponse'
      security:
      - BearerAuth: []
      summary: Sumbit speed request
      tags:
      - speedrequests
  /speedrequests/draft:
    get:
      description: Get darft infomation. Returns speedrequestid nil and count 0 for
        guest.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get darft information
      tags:
      - speedrequests
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"